reduce numpy dtypes for photons

+++ fuse/plugins/detector_physics/s1_photon_propagation.py
@@ -33,9 +33,9 @@ class S1PhotonPropagationBase(strax.Plugin):
 
     input_timeout = FUSE_PLUGIN_TIMEOUT
 
-    dtype = [('channel', np.int64),
+    dtype = [('channel', np.int16),
              ('dpe', np.bool_),
-             ('photon_gain', np.int64),
+             ('photon_gain', np.int32),
             ]
     dtype = dtype + strax.time_fields
